central_server/requirements.txt
# Core dependencies
celery>=5.2.0
redis>=4.3.0
pymongo>=4.1.0
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.9.0
docker>=6.0.0
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=0.20.0

# For admin dashboard
flask>=2.2.0
flask-socketio>=5.3.0
